P3/JWT/www/requirements.txt
flask
pycryptodome
flask-talisman